Fix cap run environment override
The cap run option -e,--environment did not work correctly when
CAP_ENV was set in a caprc file because the override was happening
before the environment was included.  This caused the override to
get overridden by environment configuration.

To fix this problem, the override was moved after the environment
configuration.