Add cap_data_download method
The cap_data_download method downloads data from a url and puts it
in the data directory.  If the file is a .tar* then the contents
will be extracted into the data directory and the original file
will be deleted.  A --md5sum option can be used to provide a sum for
the file and it will be verified after the download.