Use local `cap` command
A developer may have a version of CAPTURE installed in their system
and be developing new functionality in a different copy of the
repository elsewhere on their system. To make sure the `cap` command
in their development repository is used by tests, the development
repository directory is added to the beginning of the path in the
test suite setup.