Fix slurm error for files with parenthesis in name
When files have a parenthesis in the name and the --slurm-batch
parameter is provided, the slurm job would error. This was caused by
parenthesis in the bash script created for the slurm job. To correct the
problem, the file names in the bash script are now quoted.  This should
fix the problem for other characters too.