Fix master branch issue with new git
Old git versions would name the default branch master. This is an
insensitive name so new versions of git use main as the default.
If a user has the old version of git we rename name master to main.
If they have a new version we get an error attempting to rename a
master branch that does not exist. To resolve the issue a check for
master existing needs to be added.